Adding a new CustomColor theme option
This allows for setting arbitrary primary and secondary colors for the theme rather than being limited the material design color palates.

This also allows setting a Dark/Light palate color in a very similar API to the built in Fluent theme. This will hopefully be familiar to anyone looking to migrate from it.

This also moves the base theme properties down in the the MaterialThemeBase to allow for more generic changing between light and dark themes. This is technically a breaking change for anyone who may be rolling their own derived classes.

I also updated the demo application with commented out usages as well. This should allow for easy testing and can serve as an example for anyone looking to implement it.